# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Traffic Agentic Gateway API",
    "description": "HTTP API gateway to connect local CityFlow parsing, metrics comparison plotting, and academic RAG workflows to Dify.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://host.docker.internal:8000",
      "description": "Local Traffic Agentic Gateway"
    }
  ],
  "paths": {
    "/analyze_log": {
      "post": {
        "summary": "Extract CityFlow simulation metrics",
        "description": "Parses and extracts key traffic metrics (like ATT, throughput, etc.) from the training history log folder.",
        "operationId": "api_analyze_log_analyze_log_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeLogRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeLogResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/plot_comparison": {
      "post": {
        "summary": "Generate metrics comparison curves",
        "description": "Generates academic-quality comparison charts for vehicle throughput, average delay, queue length, and speeds, returning the HTTP URL of the image.",
        "operationId": "api_plot_comparison_plot_comparison_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlotComparisonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlotComparisonResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/search_literature": {
      "post": {
        "summary": "Retrieve literature text using local RAG",
        "description": "Searches the local PDF library for traffic signal control algorithms and methods, retrieving relevant paper paragraphs.",
        "operationId": "api_search_literature_search_literature_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchLiteratureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchLiteratureResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/search_arxiv": {
      "post": {
        "summary": "Search arXiv academic library online",
        "description": "Searches the online arXiv repository for the latest academic papers.",
        "operationId": "api_search_arxiv_search_arxiv_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchArxivRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchArxivResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/read_paper": {
      "post": {
        "summary": "Read the full content of a local paper",
        "description": "Reads and returns the complete text content of a specified local paper in the papers directory.",
        "operationId": "api_read_paper_read_paper_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadPaperRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadPaperResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/run_agent_graph": {
      "post": {
        "summary": "Run multi-agent graph evaluation workflow",
        "description": "Executes the complete multi-agent graph workflow (Analyst -> Reviewer -> Editor) to generate a full simulation evaluation or academic analysis report.",
        "operationId": "api_run_agent_graph_run_agent_graph_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunGraphRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/start_evaluation": {
      "post": {
        "summary": "Start first phase of evaluation (suspended at human review)",
        "description": "Runs the multi-agent graph from entry point (Analyst -> Reviewer) and suspends execution at the editor breakpoint.\nReturns the thread_id, parsed metrics, and generated comparison chart.",
        "operationId": "api_start_evaluation_start_evaluation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartEvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartEvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/submit_approval": {
      "post": {
        "summary": "Submit human approval/modifications to resume evaluation and compile report",
        "description": "Takes the thread_id and validated/modified metrics, updates the state in the graph database, and resumes execution to write the final academic report.",
        "operationId": "api_submit_approval_submit_approval_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalyzeLogRequest": {
        "properties": {
          "log_path": {
            "type": "string",
            "title": "Log Path",
            "description": "The local absolute path of the simulation log folder containing training_history.json"
          }
        },
        "type": "object",
        "required": [
          "log_path"
        ],
        "title": "AnalyzeLogRequest"
      },
      "AnalyzeLogResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Execution status"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result",
            "description": "The parsed traffic metrics including ATT, throughput, etc."
          }
        },
        "type": "object",
        "required": [
          "status",
          "result"
        ],
        "title": "AnalyzeLogResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PlotComparisonRequest": {
        "properties": {
          "run1_path": {
            "type": "string",
            "title": "Run1 Path",
            "description": "The local absolute path of the baseline simulation log folder"
          },
          "run2_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run2 Path",
            "description": "The local absolute path of the optimized simulation log folder (optional)"
          }
        },
        "type": "object",
        "required": [
          "run1_path"
        ],
        "title": "PlotComparisonRequest"
      },
      "PlotComparisonResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Execution status"
          },
          "image_path": {
            "type": "string",
            "title": "Image Path",
            "description": "The local absolute path of the generated image"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url",
            "description": "The HTTP URL to access the comparison image"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Status or feedback message"
          }
        },
        "type": "object",
        "required": [
          "status",
          "image_path",
          "image_url",
          "message"
        ],
        "title": "PlotComparisonResponse"
      },
      "ReadPaperRequest": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename",
            "description": "The filename of the local paper, e.g. 'AlignLight.pdf'"
          }
        },
        "type": "object",
        "required": [
          "filename"
        ],
        "title": "ReadPaperRequest"
      },
      "ReadPaperResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Execution status"
          },
          "result": {
            "type": "string",
            "title": "Result",
            "description": "The complete text content of the requested local paper"
          }
        },
        "type": "object",
        "required": [
          "status",
          "result"
        ],
        "title": "ReadPaperResponse"
      },
      "RunGraphRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "The conversational message or question to run through the entire multi-agent graph workflow"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "RunGraphRequest"
      },
      "SearchArxivRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "description": "The search query for arXiv, e.g. 'reinforcement learning traffic signal control'"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SearchArxivRequest"
      },
      "SearchArxivResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Execution status"
          },
          "result": {
            "type": "string",
            "title": "Result",
            "description": "The search results and summaries from arXiv"
          }
        },
        "type": "object",
        "required": [
          "status",
          "result"
        ],
        "title": "SearchArxivResponse"
      },
      "SearchLiteratureRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "description": "The academic query regarding traffic signal control algorithms or formulas"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SearchLiteratureRequest"
      },
      "SearchLiteratureResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Execution status"
          },
          "result": {
            "type": "string",
            "title": "Result",
            "description": "Retrieved paper paragraphs and RAG context"
          }
        },
        "type": "object",
        "required": [
          "status",
          "result"
        ],
        "title": "SearchLiteratureResponse"
      },
      "StartEvaluationRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "The conversational message or question to run through the first phase of evaluation"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "StartEvaluationRequest"
      },
      "StartEvaluationResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Execution status"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id",
            "description": "The unique thread ID of the paused evaluation run"
          },
          "is_paused": {
            "type": "boolean",
            "title": "Is Paused",
            "description": "Whether the workflow has paused at the human review breakpoint"
          },
          "baseline_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Baseline Data",
            "description": "The parsed simulation metrics of the baseline run"
          },
          "optimized_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Optimized Data",
            "description": "The parsed simulation metrics of the optimized run"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url",
            "description": "The HTTP URL of the generated metrics comparison image"
          },
          "next_node": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Next Node",
            "description": "The name of the next node in the graph, e.g., ['editor']"
          }
        },
        "type": "object",
        "required": [
          "status",
          "thread_id",
          "is_paused",
          "baseline_data",
          "optimized_data",
          "image_url",
          "next_node"
        ],
        "title": "StartEvaluationResponse"
      },
      "SubmitApprovalRequest": {
        "properties": {
          "thread_id": {
            "type": "string",
            "title": "Thread Id",
            "description": "The unique thread ID of the paused evaluation run"
          },
          "baseline_throughput": {
            "type": "string",
            "title": "Baseline Throughput",
            "description": "The validated or modified throughput of the baseline run"
          },
          "baseline_att": {
            "type": "string",
            "title": "Baseline Att",
            "description": "The validated or modified ATT of the baseline run"
          },
          "optimized_throughput": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Optimized Throughput",
            "description": "The validated or modified throughput of the optimized run"
          },
          "optimized_att": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Optimized Att",
            "description": "The validated or modified ATT of the optimized run"
          },
          "review_comment": {
            "type": "string",
            "title": "Review Comment",
            "description": "The review comment to resume the workflow with",
            "default": "Approved"
          }
        },
        "type": "object",
        "required": [
          "thread_id",
          "baseline_throughput",
          "baseline_att"
        ],
        "title": "SubmitApprovalRequest"
      },
...
    }
  }
}